feat(articles): modularize user tests
